sharepoint-qa-cli/pyproject.toml
[tool.poetry]
name = "sharepoint-qa-cli"
version = "0.1.0"
description = "A console application for extracting documents from SharePoint Online, generating embeddings, and retrieving answers via OpenAI."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
sqlalchemy = "^1.4.25"
qdrant-client = "^0.7.0"
openai = "^0.11.0"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"